Fix n / n:si conflicts
Words that would both be a noun (kelku:n) and a si-verb (kelku-n:si)
would only get one of the two shown (in practice, only the si-verb) two
times. Fixed by changing the dictionary handling to use the full type
("n" and "n:si") instead of the base type ("n" for both). For
consistency also did the same with verbs.