Be stricter in which unlenitions to accept
Previously, "mì pehilvan" would be interpreted as both "in three rivers"
(correct) and "in which river" (incorrect since that should have been
"mì fehilvan"). This commit makes the unlenition code more precise, so
that it accepts only unlenitions that would actually lenite correctly to
the query word. So, for example, "wä kxutu" would now not return any
results.