Fix incorrect word splitting in reverse search
The regex used to split up the word translations worked by splitting on
groups of \w characters, but apparently that doesn't include accented
characters. So "böse" would be split in "b" and "se", so searching for
"böse" would not find that word (searching for "b" or "se" would find
it, though...)

Thanks to the German Lerngruppe for finding this bug!