Implement a word editor
This is a big refactoring. The database is not read from navi-tsim
anymore; instead it is taken from one single JSON file words.json. This
file is then updated when edits are made. Edits are also stored in a log
file called history.json.

Edits can only be made by logged-in users, so this commit also includes
a simple authentication mechanism, which expects a JSON file called
users.json with credentials. (Note: advanced features like
user-initiated password changes are not yet supported; before this is
implemented we need mechanisms like password hashing and salting.)