Implement pseudovowel contraction for ‹er›/‹ol›
Adds a new piece of syntax to the infix specification: "(ll)" and "(rr)"
indicate that the ll/rr is unstressed and will hence disappear when ‹ol›
and ‹er› are used, respectively. Just writing "ll" and "rr" will cause
Reykunyu to assume these are stressed (and hence, the infix will
disappear instead as per Horen §2.3.2).

Fixes #6.