Show a separate affix list per conjugation step
So for example mesleleyuti gets two affix lists: one for adding -yu to
make the verb into a noun, and then another one for adding me+ and -t to
that resulting noun.

This is an API change. Results used to have an "affixes" key containing
the affix list. Now the "affixes" key is put under the "conjugated"
key instead, so that each entry of the "conjugated" array can have its
own affix list.