Make the web interface (more) multilingual
The way this works: both the server and the client side can now access
the translations (which are now stored in a JSON file). On the server
side, the EJS template no longer hard-codes strings but rather gets them
from the translations JSON file. Furthermore, it embeds the generated
translations into <span>s with the data-key attribute containing the key
from the translations file. This way, the client side can retranslate
these strings as needed when the user changes the language.

Of course all of this is kind of useless. One could just reload the page
when the user changes the language. But this is cooler, because a page
reload isn't needed.

Fixes #72.