Only remove parentheses for word links
Not for items in the sentence bar. This way, words like san ... sìk
still get their parentheses, like "(start-of-quote particle)", in the
sentence bar. But the parentheses are removed in word links, because
then we'd get double parentheses.

As a bonus, this also applies the parentheses removal to
short_translations (this was an oversight in the existing code).

Fixes #102.