Fix to-Na'vi tab still being visible in Na'vi mode
When Na'vi is selected in the language dropdown, it doesn't make sense
to search in reverse mode. There was a check for this that hid the
reverse mode option from the mode dropdown. However, this check became
broken when we moved from two separate from-Na'vi and to-Na'vi modes, to
one single mode that presents from-Na'vi and to-Na'vi tabs. In this
case, we should instead hide the tabs when Na'vi is selected in the
language dropdown. This commit fixes that oversight.

Fixes #128.