Fix result score not taking into account the query
The intention was that the result score is always 0 if the query is
equal to the root word, i.e., if no conjugation was going on. Due to not
passing the required parameter (embarrassing) it became undefined and of
course JS doesn't complain about this, hence this check simply didn't
work.

Time for TypeScript?

Fixes #122 (can be fixed even better later by looking at the number of
affixes or something, but this is fine for now).