Fix nouns ending on -ng not being found
The problem was an overzealous parsing rule. If a noun ends on -g (after
other affixes, such as -ìl, already were removed) then the rule replaced
the -g by -kx, and because that's the only possible option, I didn't
even make it bother to check the variant with -g. That doesn't work
however for nouns ending in -ng, because it would become -nkx, which
obviously doesn't make sense.

Fixed this by just also checking the unchanged variant (exactly how all
the other rules work).