Add work around for capitalization issue
Proper names would be shown in the study tool as, for example:

> Anuk (pronounced anuk)

because the pronunciation string is always lowercased.

This is still not a full solution as the pronunciation string, when it
is shown due to other (non-capitalization) differences, is still not
properly capitalized. A fix for that will have to wait until the planned
pronunciation refactor.