Return HTTP 204 to avoid XML parsing errors
An empty response is not a valid XML document, and the default content
type expected by the browser on an XMLHttpRequest is text/xml (see, for
example, https://stackoverflow.com/a/66504438). This causes an error in
the console at least in Firefox.

We're working around this by emitting HTTP 204 (No Content) instead of
HTTP 200, which silences the error.

Fixes #196.