Start porting the server-side to TypeScript
We already were using TypeScript on the client-side (using esbuild to
compile it to a JS bundle, see commit
dc411b90b5c05cad4c0e76c30048e04a789f91ac) but not yet on the
server-side.

Node doesn't natively support TS yet, so to use TS in Node, there are a
few options. One can compile the TS code to JS just like the client-side
code, but a more attractive option is to simply run the TS directly in
Node using a wrapper called tsx. That is what this commit is doing.
However, tsx doesn't actually do type checking (it simply ignores the
types) so in the build step, we now run tsc to do the type checking
(just like the client-side actually, as esbuild also doesn't do type
checking).

We're using @types/node version 16, as (1) that's the version of Node
we're currently running in production, and (2) there currently is a bug
with the latest version of @types/node that causes errors when not using
"moduleResolution": "node", whenever it is installed, even if you're not
even using any Node types. This breaks our build as we use the same
node_modules folder for the server and the client side, and on the
client side we obviously cannot use "moduleResolution": "node". See also
https://github.com/DefinitelyTyped/DefinitelyTyped/discussions/67406.