Remove unnecessary and buggy sanity checks
The sanity checks in Zeykerokyu contained three bugs:

Bug 1: The entire sanity check is unnecessary as it should never hit in
the first place. user and vocab both cannot be null given their types,
so this entire thing is pointless in the first place.

Bug 2: If the sanity check somehow (foreshadowing...) fails anyway, we
don't return after calling the callback, so the rest of the method still
gets executed. Oops! This causes us to try sending two responses for the
same HTTP request. Express (the Node.js server library we're using)
detects this, determines we're doing something stupid, and throws an
exception, which makes the Reykunyu process crash. (Usually exceptions
would trigger a HTTP 500 instead of a full-on crash. However, as we've
already sent our response as soon as this error triggers, Express can't
send another response with the HTTP 500.)

Bug 3: But if the sanity check will never fail, how is this a problem
then? Well, consider what !vocab does if vocab is a number: it evaluates
to true if vocab === 0. So which word has ID 0? Turns out that is
'angtsìk. Congratulations: everyone who tries to review the word
'angtsìk makes Reykunyu crash. Great.

This commit simply removes the sanity checks, and also makes a lot of
the other error handling more robust by simply throwing if something
goes wrong, instead of just calling the callbacks anyway and pretending
nothing was wrong.

Fixes #200.