Add service worker to allow searching when offline
This commit adds a service worker which contains the entire server-side
Reykunyu codebase bundled up into a single file. This way, the service
worker is able to intercept search API calls and run them locally in the
user's browser.

This is still very much work in progress. For now, the service worker
intercepts only Na'vi-English searches. All other API calls simply fail.
The service worker also doesn't cache resources for fully offline usage
yet.

There is a button in the settings to download dictionary data and
install the service worker. This does not properly signal
success/failure of the service worker installation yet. Also, when
searching there is no indication if the service worker was or was not
used.

For now, Zeykerokyu and some other parts of Reykunyu have been fully
commented out to make this work. Also, the editor is broken.