Replace the introduction text by the learned count
If the user has already learned words, supposedly they don't need to
read the introduction text anymore. So it makes sense to just replace it
by the learned count and the review button. (Note: the review button can
never be active if the learned count is non-zero, so it is an either/or
situation.)