Avoid spuriously opening the suggestions pane
Reykunyu used to open the pane whenever the search box gets focused.
That also happens automatically whenever the user searches for
something, though. This is inconvenient because it causes the pane to
open when the user changes search modes.

This commit makes sure the suggestions pane opens only if the user
themselves triggered the focus event, not when Reykunyu does it.

Fixes #177.