Provide LinkString references in a separate field
We used to process LinkStrings to contain TruncatedWordData objects in
the LinkString itself. This is confusing and prevents us from storing
translated LinkStrings.

Instead, we now keep the LinkStrings as they are stored in the database
(as a simple string with [word:type]-style references) and provide the
necessary references to render the LinkString in a new field
`references`.

Fixes #253.