Start a rewrite of the word editor
The old editor was buggy (#110) and didn't support translations of
fields other than the definition (#75). Also, the code was ugly so it
generally made it hard to add new features.

The new editor also shows a live preview (this needs a new API endpoint
to preprocess the edited WordData so it can be shown by the search UI
code).