Don't block headers from selecting table entries
The <h2> headers have some extra (hidden) height at the top, to ensure
that clicking one of the letters in the ToC ends up with that letter's
<h2> header under the sticky ToC bar, instead of under it.

Sadly, this hidden height stole pointer events which made it impossible
to select a cell right above the <h2> in the translations editor. (For
some reason it would still be possible to select existing text in such a
cell, but you couldn't click on the background. Strange. I have no idea
where even to begin looking for the standard that would prescribe such
behavior.)

Also, this commit updats the height of the hidden element, because in
the meantime the ToC became higher so the <h2> would still end up hidden
by the ToC.

Fixes #259.