Add the last phase to the migration script
Now there are just a lot of cases where there is a discrepancy between
the last version of a word seen by the script, and the "old" data in the
next observed edit for that word. That means that some edit happened in
between, but we apparently didn't properly process it. These are marked
as "unexplained" changes and at the moment there 4824 of them. The goal
is to get this number significantly down in the future.